+++ it.unipi.iotplatform.qosbroker.qosmanager/src/main/java/it/unipi/iotplatform/qosbroker/qosmanager/impl/QoSBrokerCore.java
@@ -1019,7 +1019,7 @@ public class QoSBrokerCore implements Ngsi10Interface, Ngsi9Interface, QoSBroker
 			
 			t.setContextEntityId(thingInfoContainer.getContextEntityId());
 			
-			Double batteryLevel = Double.valueOf(String.valueOf(contAttrsMap.get("battery").getcontextValue()));
+			Integer batteryLevel = Integer.valueOf(String.valueOf(contAttrsMap.get("battery").getcontextValue()));
 			
 			t.setBatteryLevel(batteryLevel);
